Extracted the Autocomplete to a custom component
